Add files via upload
This is a bank account app which reads data from the .csv file and display. 
if the account is checking then it creates the 13 digit card number and 4 digit pin 
if the account is savings then it creates 3 digit safety deposit box id and 4 digit key for it. 
 
this app includes the concepts of the core java and some collection concepts of the java .